# QML performance tests