--- a/frontend/qml_perf/qml_perf.go
+++ b/frontend/qml_perf/qml_perf.go
@@ -1,3 +1,7 @@
+// Copyright 2014 The lime Authors.
+// Use of this source code is governed by a 2-clause
+// BSD-style license that can be found in the LICENSE file.
+
 package main
 
 import (