--- a/frontend/qml_perf/testdata/form1.qml
+++ b/frontend/qml_perf/testdata/form1.qml
@@ -1,0 +1,7 @@
+import QtQuick 2.0
+
+Rectangle {
+    width: 100
+    height: 62
+}
+