--- a/frontend/qml_perf/testdata/form2.qml
+++ b/frontend/qml_perf/testdata/form2.qml
@@ -1,0 +1,14 @@
+import QtQuick 2.0
+import QtQuick.Controls 1.0
+import QtQuick.Controls.Styles 1.0
+import QtQuick.Dialogs 1.0
+import QtQuick.Layouts 1.0
+import QtGraphicalEffects 1.0
+//import "dialogs"
+
+
+Rectangle {
+    width: 100
+    height: 62
+}
+