--- a/frontend/qml_perf/testdata/form4.qml
+++ b/frontend/qml_perf/testdata/form4.qml
@@ -1,0 +1,7 @@
+import QtQuick 2.0
+import QtQuick.Controls 1.2
+ApplicationWindow {
+    width: 100
+    height: 62
+}
+