… frontend/qml_perf/testdata/qml_perf.qmlproject