Here is small qml performance tests. To have some understanding of performance implications of some functionality. 

The normal benchmark tests were not an option, as all qml code must run in gui thread, which benchmark tests does not atm allow you to do. 

Example of running test:
```
                                  main.QmlEngineCreationTest      100	   3449703 ns/op            3 ms/op
                        main.QmlEngineCreationAndDestroyTest      100	   4370805 ns/op            4 ms/op
                                     main.QmlEngineLoadForm1      100	      8540 ns/op
                                   main.QmlEngineCreateForm1      100	      1195 ns/op
                                     main.QmlEngineLoadForm2      100	     36401 ns/op
                                   main.QmlEngineCreateForm2      100	      1102 ns/op
                                     main.QmlEngineLoadForm3      100	    239740 ns/op
                                   main.QmlEngineCreateForm3      100	     17204 ns/op
                                     main.QmlEngineLoadForm4      100	    264181 ns/op
                                   main.QmlEngineCreateForm4      100	     19422 ns/op
                             main.QmlEngineCallingJsFunction      100	       847 ns/op
                    main.QmlEngineChangingObjectPropertyInJs      100	       773 ns/op
                    main.QmlEngineChangingObjectPropertyInGo      100	      1771 ns/op
                               main.QmlEnginePassingGoObject      100	      2042 ns/op
                  main.QmlEngineCallingGoModelFunctionFromJs      100	      1456 ns/op
     main.QmlEngineCallingGoModelFunctionFromJsWithArguments      100	      1370 ns/op

```