@pjoe 2015-01-28
Any particular reason for not using the standard golang benchmark stuff (http://golang.org/pkg/testing/)?

Also some of these are tests are somewhat artificial and won't show QML performance bottlenecks, e.g. having a single property in the root context is not an issue, but having 100 might :)