@sonicflare 2015-01-28
You have to start from something. 
The native benchmark does not allow to run tests from gui thread, and it is always run as go routine.
The go-qml have explicit checks for that, and result in panic.  

Fun fact, but the actually go-qml author created an issue about that, but just only for tests. 
https://github.com/golang/go/issues/7905