@pjoe 2015-01-28
Also these general go-qml maybe belongs better in the go-qml project?