@sonicflare 2015-01-28
It does not profile much of a go-ql code, rather than qml engine itself. 
The purpose of this, is to guide the design/refactor. To make intelligent decisions about where and how to write stuff that is interfaced with qml.
Like I don't know what thing is faster, but with this i can tell the relative performance implications of some structure in qml.