@pjoe 2015-01-28
Ok, I just meant that this didn't look very lime specific to me, but more like general go-qml benchmarks, so maybe doesn't belong too well in the lime repo as it is right now