@sonicflare 2015-01-28
That is valid point. It can go to some utility repo. 
it should be in tests, but go convention does not do that.