@pjoe 2015-01-28
Why not in go-qml? The stuff you have now seems to me to be very valuable for that project :)