@zidoms 2015-02-02
As @pjoe said these tests are related to go-qml we can't do anything about them in limetext code. We need benchmarks so we could enhance qml frontend code.