@quarnster 2015-02-05
These benchmarks would be more suitable in the [go-qml](https://github.com/go-qml/qml) repo